+++ src/Route/Segment.php
@@ -7,7 +7,7 @@
 
 declare(strict_types=1);
 
-namespace Zend\Router\Http;
+namespace Zend\Router\Route;
 
 use Traversable;
 use Zend\I18n\Translator\TranslatorInterface as Translator;
