+++ test/Route/RouteMatchTest.php
@@ -7,10 +7,10 @@
 
 declare(strict_types=1);
 
-namespace ZendTest\Router\Http;
+namespace ZendTest\Router\Route;
 
 use PHPUnit\Framework\TestCase;
-use Zend\Router\Http\RouteMatch;
+use Zend\Router\Route\RouteMatch;
 
 class RouteMatchTest extends TestCase
 {
