+++ test/Route/SchemeTest.php
@@ -7,12 +7,12 @@
 
 declare(strict_types=1);
 
-namespace ZendTest\Router\Http;
+namespace ZendTest\Router\Route;
 
 use PHPUnit\Framework\TestCase;
 use Zend\Http\Request;
-use Zend\Router\Http\RouteMatch;
-use Zend\Router\Http\Scheme;
+use Zend\Router\Route\RouteMatch;
+use Zend\Router\Route\Scheme;
 use Zend\Stdlib\Request as BaseRequest;
 use Zend\Uri\Http as HttpUri;
 use ZendTest\Router\FactoryTester;
