+++ test/Route/TestAsset/DummyRoute.php
@@ -7,10 +7,10 @@
 
 declare(strict_types=1);
 
-namespace ZendTest\Router\Http\TestAsset;
+namespace ZendTest\Router\Route\TestAsset;
 
-use Zend\Router\Http\RouteInterface;
-use Zend\Router\Http\RouteMatch;
+use Zend\Router\Route\RouteInterface;
+use Zend\Router\Route\RouteMatch;
 use Zend\Stdlib\RequestInterface;
 
 /**
